Split code into Docking_h
