+++ Demo Plugin/NppManagedPluginDemo/NppManagedPluginDemo.VS2015.csproj
@@ -48,6 +48,9 @@
     <Compile Include="..\..\Visual Studio Project Template C#\DllExport\DllExportAttribute.cs">
       <Link>DllExport\DllExportAttribute.cs</Link>
     </Compile>
+    <Compile Include="..\..\Visual Studio Project Template C#\Integration\Docking_h.cs">
+      <Link>Integration\Docking_h.cs</Link>
+    </Compile>
     <Compile Include="..\..\Visual Studio Project Template C#\NppPluginNETBase.cs">
       <Link>NppPluginNETBase.cs</Link>
     </Compile>
