+++ Visual Studio Project Template C#/$projectname$.csproj
@@ -44,6 +44,7 @@
     <Compile Include="Main.cs" />
     <Compile Include="NppPluginNETBase.cs" />
     <Compile Include="NppPluginNETHelper.cs" />
+    <Compile Include="Integration\Docking_h.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Properties\Resources.Designer.cs">
       <AutoGen>True</AutoGen>
